Remove test for `#include`ing <string>.
This fails on OS X with an error message saying that the header <string>
can't be found, and I haven't quite been able to figure out why. This
error is obviously unrelated to this change, as OS X doesn't use
`SwiftGlibc`.

It would have been nice to demonstrate that including a libc++ header
now works, but that isn't central to this change -- the important thing
about this change is that everything still works and that, additionally,
`pid_t` and similar types are now defined in the correct headers.